separate client and server funcs
